bkp/nutrition-agent/requirements.txt
langchain
langchain-community
langchain-openai
langchain-groq
groq
duckduckgo-search
pydantic
requests
python-dotenv
matplotlib
pandas
streamlit
plotly
numpy
langgraph
